Create/Edit Articles (Edit and Delete isn't working yet)

+++ NewsWeb/Controllers/TopicController.cs
@@ -14,7 +14,7 @@ namespace NewsWeb.Controllers
         private TopicManager manager = new TopicManager();
         
         [HttpGet]
-        public IActionResult Create()
+        public IActionResult Create(int? id)
         {
             if (!HttpContext.Session.GetIsAdmin())
             {
@@ -24,8 +24,18 @@ namespace NewsWeb.Controllers
             TopicModel model = new TopicModel();
             model.Topics = manager.GetAllTopics();
 
+            //in case of need to edit -> check for id
+            if (id.HasValue)
+            {
+                var data = manager.GetTopic(id.Value);
+                model.Id = data.Id;
+                model.Title = data.Title;
+            }
+            
             return View(model);
         }
+
+
         [HttpPost]
         public IActionResult Create(TopicModel model)
         {
@@ -33,10 +43,21 @@ namespace NewsWeb.Controllers
             {
                 try
                 {
-                    // manager call
-                    manager.CreateNewTopic(model.Title);
+                    // if id is not set then CREATE a NEW TOPIC
+                    if (model.Id == 0)
+                    {
+                        // manager call
+                        manager.CreateNewTopic(model.Title);
+                    }
+                    // ID is defined THAN EDIT TOPIC!!!
+                    else
+                    {
+                        manager.Update(model.Id, model.Title);
+                    }
 
-                    return RedirectToAction(nameof(Create));
+                    //redirecting back to the list of topics to see changes
+                    return RedirectToAction(nameof(Index));
+                    
                 }
                 catch (LogicException ex)
                 {
@@ -48,18 +69,78 @@ namespace NewsWeb.Controllers
                     ModelState.AddModelError("validation", ex.Message);
                 }
             }
-            
+            //if model is not valid than return to the same view. 
             return View(model);
         }
 
         public IActionResult Index()
         {
+            if (!HttpContext.Session.GetIsAdmin())
+            {
+                return NotFound();
+            }
+
             var topics = manager.GetAllTopics();
 
             return View(topics);
         }
+
+
+        [HttpGet]
+        public IActionResult Edit(int id)
+        {
+            if (!HttpContext.Session.GetIsAdmin())
+            {
+                return NotFound();
+            }
+            //information from the database
+            var data = manager.GetTopic(id); 
+
+            TopicModel model = new TopicModel();            
+            model.Topics = manager.GetAllTopics();
+
+            //filling information from DB
+            model.Title = data.Title;
+            model.Id = data.Id;
+
+
+            return View(model);
+        }
+
+        [HttpPost]
+        public IActionResult Edit(TopicModel model)
+        {
+            if (ModelState.IsValid)
+            {
+                try
+                {
+                    // need new method to call manager
+                    manager.Update(model.Id, model.Title);
+
+                    return RedirectToAction(nameof(Index));
+                }
+
+                catch (LogicException ex)
+                {
+                    ModelState.AddModelError("validation", ex.Message);
+                }
+                catch (Exception ex)
+                {
+                    // some other unexpected error
+                    ModelState.AddModelError("validation", ex.Message);
+                }
+            }
+
+            return View(model);
+        }
+
         public IActionResult Delete(int id)
         {
+            if (!HttpContext.Session.GetIsAdmin())
+            {
+                return NotFound();
+            }
+
             manager.Delete(id);
 
             return RedirectToAction(nameof(Index));
